feat: Azure Blob Storage backed impl of producer/consumer interceptors
Use topic name as container name. Random UUID for blob name.

For now we only support byte[] keys and values, to be able to determine size. We set a marker header so that the consumer knows when the payload is a claimcheck, and there for needs to be fetched externally.

Areas of future improvement:
- How the Storage account/credentials are configured, naming of config keys
- Error handling
- Support for String payload?
- More testcases
- Potential to refactor config validation/parsing

Tests:
- Azure emulator (Azurite) based IT of interceptors using testcontainers
- Complete Azurite/Kafka IT with KafkaProucer/KafkaConsumer using testcontainers

Refs:
irori-ab/claim-check-interceptors-planning#4
irori-ab/claim-check-interceptors-planning#6