feat: add ClaimCheckWrapping(De)serializers for better error handling
Interceptors will not propagate errors to producer/consumer, but serializers can.
- ClaimCheckWrappingDeserializer will replace usage of DeserializingClaimCheckConsumerInterceptor.
- SerializingClaimCheckProducerInterceptor + ClaimCheckWrappingSerializer will now produce Kafka messages
  with empty payloads, instead of null.

Remaining work is to correcly propagate some error indication from the producer interceptor to serializer.

The old configuration with just SerializingClaimCheckProducerInterceptor
+ DeserializingClaimCheckConsumerInterceptor also works, but any ClaimCheckBackend error would only
be logged as a WARN, and message still published and consumed as if the interceptor had never run.

Also:
- fix: need to manually configure wrapped serializers, not Configurable
- fix: rename Azure v8 package to avoid FQN collisions
- refactor: some method to ClaimCheckUtils